at32f4: rcc: split out common part
