+++ lib/at32/f40x/Makefile
@@ -43,7 +43,7 @@ OBJS		+= gpio.o gpio_common_all.o
 OBJS		+= i2c_common_v1.o
 OBJS		+= iwdg_common_all.o
 OBJS		+= pwr_common_v1.o
-OBJS		+= rcc.o rcc_common_all.o
+OBJS		+= rcc.o rcc_common_at32.o rcc_common_all.o
 OBJS		+= rtc.o
 OBJS		+= spi_common_all.o spi_common_v1.o
 OBJS		+= timer.o timer_common_all.o
@@ -55,6 +55,6 @@ OBJS		+= usb_audio.o usb_cdc.o usb_midi.o
 OBJS		+= st_usbfs_core.o st_usbfs_v1.o
 
 vpath %.c	../../stm32/f1:../../stm32/common:../../stm32
-VPATH 		+= ../../usb:../../cm3:../../ethernet
+VPATH 		+= ../common:../../usb:../../cm3:../../ethernet
 
 include ../../Makefile.include
